requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
stripe
python-dotenv
fpdf
qrcode[pil]